+++ src/dashboard/routes.ts
@@ -1,6 +1,7 @@
 import { GlobalContext, LocalContext } from "../types/context"
 import parsePath from "../functions/parsePath"
 import { getFilesRecursively } from "rjutils-collection"
+import { getPreviousHours } from "../classes/dataStat"
 import { RequestContext } from "../types/external"
 import WebSocket from "../types/webSocket"
 import HTTP from "../types/http"
@@ -74,6 +75,7 @@ export default async function statsRoute(ctr: RequestContext, ctg: GlobalContext
           const date = new Date()
           const startTime = date.getTime()
           const startUsage = process.cpuUsage()
+          const previousHours = getPreviousHours()
   
           const staticFiles = await new Promise<number>(async(resolve) => {
             let staticFiles = 0
@@ -96,31 +98,31 @@ export default async function statsRoute(ctr: RequestContext, ctg: GlobalContext
               if (index === 0) return ctg.requests.stats.total
               else if (index === 1) return ctg.requests.stats.perSecond
               else return {
-                hour: ctx.previousHours[index - 2],
-                amount: ctg.requests.stats[ctx.previousHours[index - 2]]
+                hour: previousHours[index - 2],
+                amount: ctg.requests.stats[previousHours[index - 2]]
               }
             }), web_sockets: {
               opened: Array.from({ length: 7 }, (value, index) => {
                 if (index === 0) return ctg.webSockets.opened.stats.total
                 else if (index === 1) return ctg.webSockets.opened.stats.perSecond
                 else return {
-                  hour: ctx.previousHours[index - 2],
-                  amount: ctg.webSockets.opened.stats[ctx.previousHours[index - 2]]
+                  hour: previousHours[index - 2],
+                  amount: ctg.webSockets.opened.stats[previousHours[index - 2]]
                 }
               }), messages: {
                 incoming: Array.from({ length: 7 }, (value, index) => {
                   if (index === 0) return ctg.webSockets.messages.incoming.stats.total
                   else if (index === 1) return ctg.webSockets.messages.incoming.stats.perSecond
                   else return {
-                    hour: ctx.previousHours[index - 2],
-                    amount: ctg.webSockets.messages.incoming.stats[ctx.previousHours[index - 2]]
+                    hour: previousHours[index - 2],
+                    amount: ctg.webSockets.messages.incoming.stats[previousHours[index - 2]]
                   }
                 }), outgoing: Array.from({ length: 7 }, (value, index) => {
                   if (index === 0) return ctg.webSockets.messages.outgoing.stats.total
                   else if (index === 1) return ctg.webSockets.messages.outgoing.stats.perSecond
                   else return {
-                    hour: ctx.previousHours[index - 2],
-                    amount: ctg.webSockets.messages.outgoing.stats[ctx.previousHours[index - 2]]
+                    hour: previousHours[index - 2],
+                    amount: ctg.webSockets.messages.outgoing.stats[previousHours[index - 2]]
                   }
                 })
               }
@@ -129,15 +131,15 @@ export default async function statsRoute(ctr: RequestContext, ctg: GlobalContext
                 if (index === 0) return ctg.data.incoming.stats.total
                 else if (index === 1) return ctg.data.incoming.stats.perSecond
                 else return {
-                  hour: ctx.previousHours[index - 2],
-                  amount: ctg.data.incoming.stats[ctx.previousHours[index - 2]]
+                  hour: previousHours[index - 2],
+                  amount: ctg.data.incoming.stats[previousHours[index - 2]]
                 }
               }), outgoing: Array.from({ length: 7 }, (value, index) => {
                 if (index === 0) return ctg.data.outgoing.stats.total
                 else if (index === 1) return ctg.data.outgoing.stats.perSecond
                 else return {
-                  hour: ctx.previousHours[index - 2],
-                  amount: ctg.data.outgoing.stats[ctx.previousHours[index - 2]]
+                  hour: previousHours[index - 2],
+                  amount: ctg.data.outgoing.stats[previousHours[index - 2]]
                 }
               })
             }, cpu: {